requirements.txt
django==5.1
djangorestframework==3.15.2
python-telegram-bot==13
urllib3==1.26.15
psycopg2==2.9.9
django-redis==5.4.0